refactor: split checklists and rotulometro components
